# Project/Project/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Student Evaluator System API
  description: |
    The Student Evaluator System provides automated evaluation of Java programming assignments through JUnit testing.
    This API allows you to upload assignments, manage test cases, trigger evaluations, and retrieve results.
  version: 1.0.0
  contact:
    name: Student Evaluator System
    email: support@studentevaluator.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.studentevaluator.com/v1
    description: Production server
paths:
...
  /students/{studentId}/scores:
    get:
      summary: Get student scores
      description: Retrieve all scores for a specific student
      operationId: getStudentScores
      parameters:
        - name: studentId
          in: path
          required: true
          description: Student identifier
          schema:
            type: string
      responses:
        '200':
          description: Student scores retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentScoreResponse'
        '404':
          description: Student not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Students
  /students/{studentId}/evaluations:
    get:
      summary: Get student evaluation history
      description: Retrieve evaluation history for a specific student
      operationId: getStudentEvaluations
      parameters:
        - name: studentId
          in: path
          required: true
          description: Student identifier
          schema:
            type: string
        - name: assignmentId
          in: query
          description: Filter by assignment ID
          schema:
            type: string
        - name: status
          in: query
          description: Filter by evaluation status
          schema:
            type: string
            enum: [PENDING, COMPLETED, FAILED]
      responses:
        '200':
          description: Student evaluation history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EvaluationResponse'
        '404':
          description: Student not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      tags:
        - Students
components:
  schemas:
...
    EvaluationResponse:
      type: object
      properties:
        evaluationId:
          type: string
          description: Unique evaluation identifier
          example: "eval-123-456"
        studentId:
          type: string
          description: Student identifier
          example: "student123"
        assignmentId:
          type: string
          description: Assignment identifier
          example: "calc-assignment-1"
        status:
          type: string
          enum: [PENDING, COMPLETED, FAILED]
          description: Current evaluation status
        score:
          type: number
          format: double
          description: Evaluation score
          example: 85.5
        maxScore:
          type: number
          format: double
          description: Maximum possible score
          example: 100.0
        evaluatedAt:
          type: string
          format: date-time
          description: Evaluation completion timestamp
      required:
        - evaluationId
        - studentId
        - assignmentId
        - status
...
    StudentScoreResponse:
      type: object
      properties:
        studentId:
          type: string
          description: Student identifier
        name:
          type: string
          description: Student name
        scores:
          type: array
          items:
            type: object
            properties:
              assignmentId:
                type: string
              assignmentTitle:
                type: string
              score:
                type: number
                format: double
              maxScore:
                type: number
                format: double
              evaluatedAt:
                type: string
                format: date-time
        averageScore:
          type: number
          format: double
          description: Average score across all assignments
      required:
        - studentId
        - scores
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Error code
              example: "COMPILATION_FAILED"
            message:
              type: string
              description: Human-readable error message
              example: "Java compilation failed"
            details:
              type: object
              description: Additional error details
            timestamp:
              type: string
              format: date-time
              description: Error timestamp
          required:
            - code
            - message
            - timestamp
      required:
        - error
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (future enhancement)
...